+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # REPLACE-GLOBAL-WITH-NEXT-VERSION
+- PT-10448 - Adds onboarding to get API credentials with PayPal login in settings module
 - PT-11292 - Adds possibility to enter separate credentials for sandbox mode in first run wizard
 - PT-11498 - Adds PayPal Express Button to QuickView from CMS Extensions plugin